+++ src/DataTypes/Common/SPRITE/Sprite.cs
@@ -108,8 +108,8 @@ namespace BinarySerializer.Ray1
                 
                 s.DoBits<ushort>(b =>
                 {
-                    JAG_Byte_05 = (byte)b.SerializeBits<int>(JAG_Byte_05, 6, name: nameof(JAG_Byte_05));
-                    Height = (ushort)b.SerializeBits<int>(Height, 10, name: nameof(Height));
+                    JAG_Byte_05 = b.SerializeBits<ushort>(JAG_Byte_05, 6, name: nameof(JAG_Byte_05));
+                    Height = b.SerializeBits<ushort>(Height, 10, name: nameof(Height));
                 });
 
                 JAG_Ushort_07 = s.Serialize<ushort>(JAG_Ushort_07, name: nameof(JAG_Ushort_07));
@@ -135,16 +135,16 @@ namespace BinarySerializer.Ray1
                         HitBoxHeight = s.Serialize<byte>(HitBoxHeight, name: nameof(HitBoxHeight));
                         s.DoBits<byte>(b =>
                         {
-                            HitBoxOffsetX = (byte)b.SerializeBits<int>(HitBoxOffsetX, 4, name: nameof(HitBoxOffsetX));
-                            HitBoxOffsetY = (byte)b.SerializeBits<int>(HitBoxOffsetY, 4, name: nameof(HitBoxOffsetY));
+                            HitBoxOffsetX = b.SerializeBits<byte>(HitBoxOffsetX, 4, name: nameof(HitBoxOffsetX));
+                            HitBoxOffsetY = b.SerializeBits<byte>(HitBoxOffsetY, 4, name: nameof(HitBoxOffsetY));
                         });
                         Unknown4 = s.Serialize<byte>(Unknown4, name: nameof(Unknown4));
                         ImageOffsetInPageX = s.Serialize<byte>(ImageOffsetInPageX, name: nameof(ImageOffsetInPageX));
                         ImageOffsetInPageY = s.Serialize<byte>(ImageOffsetInPageY, name: nameof(ImageOffsetInPageY));
                         s.DoBits<ushort>(b =>
                         {
-                            PaletteX = (ushort)b.SerializeBits<int>(PaletteX, 6, name: nameof(PaletteX));
-                            PaletteY = (ushort)b.SerializeBits<int>(PaletteY, 10, name: nameof(PaletteY));
+                            PaletteX = b.SerializeBits<ushort>(PaletteX, 6, name: nameof(PaletteX));
+                            PaletteY = b.SerializeBits<ushort>(PaletteY, 10, name: nameof(PaletteY));
                         });
                         TexturePageInfo = s.SerializeObject<PS1_TSB>(TexturePageInfo, name: nameof(TexturePageInfo));
                     }
@@ -184,8 +184,8 @@ namespace BinarySerializer.Ray1
                         HitBoxHeight = s.Serialize<byte>(HitBoxHeight, name: nameof(HitBoxHeight));
                         s.DoBits<byte>(b =>
                         {
-                            HitBoxOffsetX = (byte)b.SerializeBits<int>(HitBoxOffsetX, 4, name: nameof(HitBoxOffsetX));
-                            HitBoxOffsetY = (byte)b.SerializeBits<int>(HitBoxOffsetY, 4, name: nameof(HitBoxOffsetY));
+                            HitBoxOffsetX = b.SerializeBits<byte>(HitBoxOffsetX, 4, name: nameof(HitBoxOffsetX));
+                            HitBoxOffsetY = b.SerializeBits<byte>(HitBoxOffsetY, 4, name: nameof(HitBoxOffsetY));
                         });
                     }
 
@@ -196,8 +196,8 @@ namespace BinarySerializer.Ray1
                         Unknown4 = s.Serialize<byte>(Unknown4, name: nameof(Unknown4));
                         s.DoBits<ushort>(b =>
                         {
-                            PaletteX = (ushort)b.SerializeBits<int>(PaletteX, 6, name: nameof(PaletteX));
-                            PaletteY = (ushort)b.SerializeBits<int>(PaletteY, 10, name: nameof(PaletteY));
+                            PaletteX = b.SerializeBits<ushort>(PaletteX, 6, name: nameof(PaletteX));
+                            PaletteY = b.SerializeBits<ushort>(PaletteY, 10, name: nameof(PaletteY));
                         });
                         TexturePageInfo = s.SerializeObject<PS1_TSB>(TexturePageInfo, name: nameof(TexturePageInfo));
                         ImageOffsetInPageX = s.Serialize<byte>(ImageOffsetInPageX, name: nameof(ImageOffsetInPageX));
@@ -215,8 +215,8 @@ namespace BinarySerializer.Ray1
                     HitBoxHeight = s.Serialize<byte>(HitBoxHeight, name: nameof(HitBoxHeight));
                     s.DoBits<byte>(b =>
                     {
-                        HitBoxOffsetX = (byte)b.SerializeBits<int>(HitBoxOffsetX, 4, name: nameof(HitBoxOffsetX));
-                        HitBoxOffsetY = (byte)b.SerializeBits<int>(HitBoxOffsetY, 4, name: nameof(HitBoxOffsetY));
+                        HitBoxOffsetX = b.SerializeBits<byte>(HitBoxOffsetX, 4, name: nameof(HitBoxOffsetX));
+                        HitBoxOffsetY = b.SerializeBits<byte>(HitBoxOffsetY, 4, name: nameof(HitBoxOffsetY));
                     });
                     Unknown3 = s.Serialize<byte>(Unknown3, name: nameof(Unknown3));
                     Unknown4 = s.Serialize<byte>(Unknown4, name: nameof(Unknown4));
